Simplify async option loading API to hook only
