+++ packages/react-querybuilder/src/components/barrel.ts
@@ -12,4 +12,3 @@ export * from './RuleGroup';
 export * from './ShiftActions';
 export * from './ValueEditor';
 export * from './ValueSelector';
-export * from './ValueSelectorAsync';
